Refactor trail path utilities and enhance trail settings
- Introduced `trail-path-utils.ts` for trail path operations including smoothing, trimming, and fallback generation.
- Updated `control-panel-settings.tsx` to increase maximum trail distance from 100 to 120.
- Adjusted polling interval in `use-flights.ts` from 10s to 5s for improved data density.
- Increased maximum trail distance in `use-settings.tsx` from 100 to 120 and updated default trail distance from 40 to 80.
- Expanded maximum points in `use-trail-history.ts` from 55 to 120.
- Modified maximum altitude for departure detection in `route-detection.ts` from 3000m to 1500m.
- Changed cache description in `route-lookup.ts` from LRU to FIFO.
- Implemented post-processing for stitched trail paths in `trail-stitch-postprocess.ts` to ensure trails reach aircraft position, filter NaN/Infinity coordinates, cap total path length, and smooth junctions.
- Refactored `stitchHistoricalTrail` in `trail-stitching.ts` to utilize new post-processing functions for improved trail quality.
