+++ src/components/map/flight-math.ts
@@ -1,0 +1,31 @@
+import type { Snapshot } from "./flight-layer-constants";
+
+// ── Interpolation Math ─────────────────────────────────────────────────
+
+export function lerpAngle(a: number, b: number, t: number): number {
+  const delta = ((b - a + 540) % 360) - 180;
+  return a + delta * t;
+}
+
+export function smoothStep(t: number): number {
+  return t * t * (3 - 2 * t);
+}
+
+// ── Distance Helpers ───────────────────────────────────────────────────
+
+export function horizontalDistanceFromLngLat(
+  aLng: number,
+  aLat: number,
+  bLng: number,
+  bLat: number,
+): number {
+  const avgLatRad = ((aLat + bLat) * 0.5 * Math.PI) / 180;
+  const metersPerDegLon = 111_320 * Math.max(0.2, Math.cos(avgLatRad));
+  const dx = (bLng - aLng) * metersPerDegLon;
+  const dy = (bLat - aLat) * 111_320;
+  return Math.hypot(dx, dy);
+}
+
+export function horizontalDistanceMeters(a: Snapshot, b: Snapshot): number {
+  return horizontalDistanceFromLngLat(a.lng, a.lat, b.lng, b.lat);
+}
